Add files via upload
Cleaned up version of the program. Removed commented out sections of the code, and added comments to help in understanding the functionality of the different parts in the program. Ran last checks to make sure that the program compiled properly and produced results that were logical given the flow of the program.