Add files via upload
Adding updated final code, and the final design document. Unit Tests are in an unmerged branch so that we did not run the risk of corrupting or overriding the project's master branch files.
Also changed from string.h to string and stdlib.h to cstdlib in base.h to allow the project to work correctly in Visual Studios.